+++ module/canvas/tokenruler.js
@@ -24,12 +24,15 @@ export default class TorgEternityTokenRuler extends foundry.canvas.placeables.to
   _getGridHighlightStyle(waypoint, offset) {
     const result = super._getGridHighlightStyle(waypoint, offset);
     if (waypoint.action !== 'walk') return result;
-    const move = this.token?.actor.system.other.move;
+    const move = this.token.actor?.system?.other?.move;
+    if (!move) return result;
     const cost = waypoint.measurement.cost;
     if (cost > 3 * move)
-      result.color = 0xce0707;  // color-level-error
+      result.color = game.settings.get('torgeternity', 'rulerGridMax');  // color-level-error
     else if (cost > move)
-      result.color = 0xee9b3a;  // color-level-warning
+      result.color = game.settings.get('torgeternity', 'rulerGridRun');  // color-level-warning
+    else
+      result.color = game.settings.get('torgeternity', 'rulerGridWalk');  // color-level-warning
     return result;
   }
 }