+++ Changelog.md
@@ -5,6 +5,7 @@
 - Scene Navigation returned to Torg styling.
 - Add option to disable fading of Scene Navigation menu.
 - Combat Tracker no longer disappears when a card Hand is deleted (which would occur when an Actor is deleted).
+- Token Ruler grid highlight is coloured according to move/max move of the token (if any).
 
 ## v13.2.2